feat(admin): event feed of all users exposed in admin

+++ lib/api/payments.ts
@@ -42,7 +42,7 @@ export async function listPayments({
     return {
       id: dbPayment.id,
       status: dbPayment.status,
-      amount: parseInt(dbPayment.amount),
+      amount: parseFloat(dbPayment.amount),
       token: dbPayment.token,
       user: getPaymentUser(dbPayment),
       date: dbPayment.updatedAt,
